+++ src/components/navigation/nav-menu/components/nav-icon-item/index.ts
