+++ src/components/navigation/nav-menu/nav-menu.types.ts
@@ -1,6 +1,6 @@
-import { NavDropdownItemProps } from './items/nav-drop-down-item/nav-drop-down-item.types';
-import { NavIconItemProps } from './items/nav-icon-item';
-import { NavItemProps } from './items/nav-item';
+import { NavDropdownItemProps } from './components/nav-drop-down-item/nav-drop-down-item.types';
+import { NavIconItemProps } from './components/nav-icon-item';
+import { NavItemProps } from './components/nav-item';
 
 export type NavMenuProps = {
 	/** Logo rendered as a linked icon on the left side of the nav */
